requirements.txt
streamlit
torch
torchvision
transformers
scikit-learn
Pillow
numpy
pandas
statsmodels
wandb